pci: Translate PCI addresses to host addresses at the end of map-in
Currently, it is not possible to use VGA devices attached to a
PCI bridge on board-qemu, e.g. by starting QEMU like this:

 qemu-system-ppc64 -nodefaults -device pci-bridge,id=br1,chassis_nr=1 \
        -serial mon:stdio -device VGA,id=video,bus=br1,addr=1

One of the problems is the missing translate-address at the end
of the map-in function of the bridge - which was already marked
as a TODO, but apparently has never been enabled. So let's do
that now!

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>