usb/storage: Implement block write support
The only missing parts were to manage the transfer direction in
do-bulk-command and to copy the data to the buffer before the
write operation.

This is needed as GRUB2 wants to write the grubenv file at start
and hangs because the data are not provided to the disk controller.

I've checked the file is correctly modified by modifying an environment
variable in GRUB2 with "set saved_entry=2" then "save_env saved_entry"
and checking the result in linux with "grub2-editenv list".

Fixes: Fixes: a0b96fe66fcd991b407c1d67ca842921e477a6fd
       (Provide "write" function in the disk-label package)
Signed-off-by: Laurent Vivier <lvivier@redhat.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>