ibm,client-architecture-support: Fix stack handling
fdt-fix-cas-node returns the end address after it's finished which
the caller (ibm,client-architecture-support) does not use or drop.
This renames fdt-fix-cas-node to (fdt-fix-cas-node) and adds a wrapper
on top of that which does the drop. This will be used later for 2-pass
tree patching.

While at this, exit the function if memory allocation failed.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>