js2x: Fix compile and cleanup
This moves shared symbols out from the header file and leaves only
the minimum in bmc.h.

This fixes a "while (...); {" bug spotted by the modern gcc.

This fixes cleanup as otherwise the board-js2x/Makefile's "clean" tries
cleaning "rtas" which tries cleaning all libs (not just the ones it uses)
and that in turn triggers generation of lib/libtpm/Makefile.dep which
fails due to missing ./board-js2x/include/version.h.

This fixes create_reloc_table.sh to build with newer toolchains which
otherwise fail with:

ld: cannot use executable file 'reloc_table.o' as input to a link

This adds js2x compile to the travis script to know when it breaks
again.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>