tpm: Implement firmware API call pass-through-to-tpm
Implement the firmware API call pass-through-to-tpm that allows a caller
to pass a TPM command to the TPM. Since the buffer provided by the user
will be used for returning the TPM's response it must be sufficiently
large. To be safe, it should be of the size returned by the firmware API
call tpm-get-maximum-cmd-size.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
[aik: replaced memcpy(hdr->totlen) with direct access]
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>